Address PR request issues
Some code was moved to llvmoExpose.cc - I'm not
sure why I had to do this now to get things to work.
