+++ include/clasp/llvmo/llvmoPackage.h
@@ -65,6 +65,7 @@ bool llvm_sys__load_bc(core::Pathname_sp filename, bool verbose, bool print, cor
 
 ClaspJIT_sp llvm_sys__clasp_jit();
 
+
 void initialize_llvm();
 void initialize_ClaspJIT();
